refactor(story): isolate internal schema helpers
