+++ desktop/src/assets/assetProtocol.test.ts
