+++ desktop/src/assets/assetProtocol.ts
