+++ desktop/src/assets/localAssetOpen.test.ts
