+++ desktop/src/assets/localAssetOpen.ts
