+++ desktop/src/assets/localAssetPolicy.test.ts
