+++ desktop/src/assets/localAssetPolicy.ts
