+++ desktop/src/assets/localAssetRegistry.test.ts
