+++ desktop/src/assets/localAssetRegistry.ts
