+++ desktop/src/assets/preloadLocalAssetCache.test.ts
