+++ desktop/src/assets/preloadLocalAssetCache.ts
