+++ desktop/src/bridge/bridgeClient.test.ts
