+++ desktop/src/bridge/bridgeLifecycle.ts
