+++ desktop/src/bridge/shared.ts
