+++ desktop_bridge/voice/voice_assets.py
