+++ story_simulation/internal/__init__.py
@@ -1,0 +1,5 @@
+"""Implementation details for the Story bounded context.
+
+Modules in this package are intentionally not part of the public Story API.
+"""
+
