+++ tests/desktop_bridge/voice/test_voice_assets.py
